pyproject.toml
[project]
name = "speaker-retrieval"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    # Core
    "numpy>=1.26",
    "pandas>=2.2",
    "scipy>=1.13",
    "scikit-learn>=1.5",
    "matplotlib>=3.9",
    "tqdm>=4.66",
    "pyyaml>=6.0",
    "omegaconf>=2.3",
    # Audio
    "soundfile>=0.12",
    "librosa>=0.10",
    # ONNX inference
    "onnxruntime-gpu>=1.19",
    "onnx>=1.16",
    # Deep learning
    "torch>=2.4",
    "torchaudio>=2.4",
    "transformers>=4.44",
    "huggingface-hub>=0.24",
    "speechbrain>=1.0",
    # Retrieval
    "faiss-gpu-cu12>=1.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
run-baseline = "scripts.run_baseline:main"
run-eval = "scripts.run_eval:main"
